수정:6/26 19:26 배포직전수정안

+++ src/main/java/point/dto/PointDTO.java
@@ -9,6 +9,6 @@ import java.time.LocalDate;
 @Setter
 public class PointDTO {
     private String empId;
-    private int pointsEarned;
+    private double pointsEarned;
     private LocalDate date;
 }
